+++ desktop/modal/icon.selector.php
@@ -16,7 +16,7 @@
 */
 
 if (!isConnect()) {
-  throw new Exception('{{401 - Accès non autorisé}}');
+  throw new Exception('401 - {{Accès non autorisé}}');
 }
 
 $objectId = init('object_id');
@@ -29,7 +29,6 @@ sendVarToJS([
 ?>
 
 <div style="display: none;" id="div_iconSelectorAlert" data-modalType="md_iconSelector"></div>
-
 <ul class="nav nav-tabs" role="tablist">
   <?php if (!$objectId) { ?>
     <li role="presentation" class="active"><a href="#tabicon" aria-controls="home" role="tab" data-toggle="tab"><i class="fas fa-icons"></i> {{Icônes}}</a></li>
@@ -76,19 +75,16 @@ sendVarToJS([
                 echo '<li><a data-path="' . $root . '/' . $dir . '">' . ucfirst(str_replace(array('/', '_'), array('', ' '), $dir)) . '</a></li>';
               }
             }
-            echo '<li><a data-path="'.__DIR__ . '/../../3rdparty/font-awesome5/">Font-Awesome</a></li>';
+            echo '<li><a data-path="' . __DIR__ . '/../../3rdparty/font-awesome5/">Font-Awesome</a></li>';
             ?>
           </ul>
         </div>
-        <div class="div_imageGallery">
-        </div>
+        <div class="div_imageGallery"></div>
       </div>
     </div>
 
     <div role="tabpanel" class="tab-pane" id="tabimg">
-      <span class="btn btn-default btn-file" style="position:absolute;right:25px;top:40px;">
-        <i class="fas fa-cloud-upload-alt"></i> {{Envoyer}}<input id="bt_uploadImg" type="file" name="file" multiple="multiple" data-path="" style="display: inline-block;">
-      </span>
+      <input class="hidden" id="bt_uploadImg" type="file" name="file" multiple="multiple" data-path="">
       <div class="imgContainer">
         <div id="treeFolder-img" class="div_treeFolder">
           <ul id="ul_Folder-img">
@@ -100,14 +96,14 @@ sendVarToJS([
             ?>
           </ul>
         </div>
-        <div class="div_imageGallery">
-        </div>
+        <div class="div_imageGallery"></div>
       </div>
     </div>
   <?php } else { ?>
     <div role="tabpanel" class="tab-pane active" id="tabobjectbg">
       <div class="imgContainer">
         <div id="treeFolder-bg" class="div_treeFolder">
+          <sup class="pull-right"><i class="fas fa-question-circle" title="{{Clic droit sur un dossier pour ouvrir le menu contextuel}}"></i></sup>
           <ul id="ul_Folder-bg">
             <?php
             $rootPath = __DIR__ . '/../../core/img/object_background/';
@@ -117,8 +113,7 @@ sendVarToJS([
             ?>
           </ul>
         </div>
-        <div class="div_imageGallery">
-        </div>
+        <div class="div_imageGallery"></div>
       </div>
     </div>
   <?php } ?>
@@ -131,13 +126,15 @@ sendVarToJS([
     buttonSet.find('#mySearch').remove()
     var mySearch = $('.ui-dialog[aria-describedby="mod_selectIcon"]').find('#mySearch')
     buttonSet.append(mySearch)
+
+    $('#treeFolder-img').prepend('<sup class="pull-right"><i class="fas fa-question-circle" title="{{Clic droit sur un dossier pour ouvrir le menu contextuel}}"></i></sup>')
   })
 
   $('.div_treeFolder').off('click').on('select_node.jstree', function(node, selected) {
     $('#in_searchIconSelector').val('')
     if (selected.node.a_attr['data-path'] != undefined) {
-      var path = selected.node.a_attr['data-path'];
-      printFileFolder(path, $(this).attr('id'));
+      var path = selected.node.a_attr['data-path']
+      printFileFolder(path, $(this).attr('id'))
       if ($(this).attr('id') === 'treeFolder-img') {
         var ref = $(this).jstree(true)
         var sel = ref.get_selected()[0]
@@ -173,13 +170,122 @@ sendVarToJS([
         })
       }
     }
+  }).on("rename_node.jstree", function(event, data) {
+    if (data.text !== data.old) {
+      var newPath = data.node.a_attr['data-path'].replace('/' + data.old + '/', '/' + data.text + '/')
+      jeedom.renameFolder({
+        src: data.node.a_attr['data-path'],
+        dst: newPath,
+        error: function(error) {
+          $.fn.showAlert({
+            message: error.message,
+            level: 'danger'
+          })
+        },
+        success: function(result) {
+          if (result) {
+            $.jstree.reference(data.node).get_node(data.node.id).a_attr['data-path'] = newPath
+            $('#bt_uploadImg').attr('data-path', newPath.substr(newPath.search('data/')))
+          }
+        }
+      })
+    }
   })
 
   $(".div_treeFolder").jstree({
     "core": {
-      "check_callback": true
-    }
-  });
+      "check_callback": true,
+      "force_text": true,
+      "themes": {
+        "icons": false
+      }
+    },
+    "contextmenu": {
+      "items": function(node) {
+        var tree = $.jstree.reference(node)
+        if (tree.element[0].id !== 'treeFolder-img') {
+          return false
+        }
+        var items = {
+          "upload": {
+            "separator_after": true,
+            "label": "{{Envoyer des images}}",
+            "icon": "fas fa-file-upload",
+            "action": function() {
+              $('#bt_uploadImg').click()
+            }
+          },
+          "create": {
+            "label": "{{Nouveau dossier}}",
+            "icon": "fas fa-folder-plus",
+            "action": function() {
+              jeedom.createFolder({
+                name: '{{Nouveau}}',
+                path: node.a_attr['data-path'],
+                error: function(error) {
+                  $.fn.showAlert({
+                    message: error.message,
+                    level: 'danger'
+                  })
+                },
+                success: function() {
+                  tree.create_node(node, {
+                    text: "{{Nouveau}}",
+                    a_attr: {
+                      "data-path": node.a_attr['data-path'] + '{{Nouveau}}/'
+                    }
+                  }, 'inside', function(newNode) {
+                    tree.deselect_node(node)
+                    tree.select_node(newNode)
+                    tree.edit(newNode)
+                  })
+                }
+              })
+            }
+          }
+        }
+        if (node.parent !== '#') {
+          items.rename = {
+            "label": "{{Renommer le dossier}}",
+            "icon": "fas fa-folder",
+            "action": function() {
+              tree.edit(node)
+            }
+          }
+          items.remove = {
+            "label": "{{Supprimer le dossier}}",
+            "icon": "fas fa-folder-minus",
+            "action": function() {
+              bootbox.confirm("{{Etes-vous sûr de vouloir supprimer le dossier}} <strong>" + tree.get_node(node.parent).text + "/" + node.text + "</strong> ?<br>{{Attention : le contenu du dossier sera définitivement supprimé lors de l'opération.}}", function(result) {
+                if (result) {
+                  jeedom.deleteFolder({
+                    path: node.a_attr['data-path'],
+                    error: function(error) {
+                      $.fn.showAlert({
+                        message: error.message,
+                        level: 'danger'
+                      })
+                    },
+                    success: function(result) {
+                      if (result) {
+                        tree.delete_node(node)
+                        tree.select_node(node.parent)
+                      }
+                    }
+                  })
+                }
+              })
+            }
+          }
+        }
+        return items
+      }
+    },
+    "unique": {
+      "case_sensitive": true
+    },
+    "plugins": ["contextmenu", "unique"]
+  })
 
   function printFileFolder(_path, jstreeId) {
     $.hideAlert()
@@ -193,66 +299,66 @@ sendVarToJS([
         })
       },
       success: function(data) {
-        var div = '';
-        $('#' + jstreeId).siblings('.div_imageGallery').empty();
+        var div = ''
+        $('#' + jstreeId).siblings('.div_imageGallery').empty()
         if (_path.indexOf('data/') != -1) {
-          var realPath = _path.substr(_path.search('data/'));
+          var realPath = _path.substr(_path.search('data/'))
         } else if (_path.indexOf('core/') != -1) {
-          var realPath = _path.substr(_path.search('core/'));
+          var realPath = _path.substr(_path.search('core/'))
         } else {
-          var realPath = _path.substr(_path.search('3rdparty/'));
+          var realPath = _path.substr(_path.search('3rdparty/'))
         }
         if (jstreeId === 'treeFolder-icon') {
           $('#sel_colorIcon').show()
-          var category = realPath.slice(0, -1).split('/').pop();
+          var category = realPath.slice(0, -1).split('/').pop()
           if (category === 'font-awesome5') {
             jeedom.getFileContent({
               path: realPath + 'icons.json',
               success: function(data) {
-                data = JSON.parse(data);
+                data = JSON.parse(data)
                 for (var i in data.icons) {
                   var test = (iconClasses && iconClasses[2] === data.icons[i].substr(4)) ? ' iconSelected' : ''
-                  div += '<div class="divIconSel text-center' + test + '">';
-                  div += '<span class="cursor iconSel"><i class="' + data.icons[i] + ' ' + $('#sel_colorIcon').value() + '"></i></span><br/><span class="iconDesc">' + data.icons[i].substr(7) + '</span>';
-                  div += '</div>';
+                  div += '<div class="divIconSel text-center' + test + '">'
+                  div += '<span class="cursor iconSel"><i class="' + data.icons[i] + ' ' + $('#sel_colorIcon').value() + '"></i></span><br/><span class="iconDesc">' + data.icons[i].substr(7) + '</span>'
+                  div += '</div>'
                 }
                 $('#' + jstreeId).siblings('.div_imageGallery').append(div)
 
               }
-            });
+            })
 
           } else {
             $.get(realPath + 'style.css', function(data, status) {
               if (status === 'success') {
-                var exp_reg = new RegExp('(?=\.)' + category + '-(.*?).+?(?=\:)', "gi");
-                var matches = data.match(exp_reg);
+                var exp_reg = new RegExp('(?=\.)' + category + '-(.*?).+?(?=\:)', "gi")
+                var matches = data.match(exp_reg)
                 for (var i in matches) {
                   var selected = (iconClasses && iconClasses[2] === matches[i]) ? ' iconSelected' : ''
-                  div += '<div class="divIconSel text-center' + selected + '">';
-                  div += '<span class="cursor iconSel"><i class=\'icon ' + matches[i] + ' ' + $('#sel_colorIcon').value() + '\'></i></span><br/><span class="iconDesc">' + matches[i].replace(category + '-', '') + '</span>';
-                  div += '</div>';
+                  div += '<div class="divIconSel text-center' + selected + '">'
+                  div += '<span class="cursor iconSel"><i class=\'icon ' + matches[i] + ' ' + $('#sel_colorIcon').value() + '\'></i></span><br/><span class="iconDesc">' + matches[i].replace(category + '-', '') + '</span>'
+                  div += '</div>'
                 }
                 $('#' + jstreeId).siblings('.div_imageGallery').append(div)
               }
-            });
+            })
           }
         } else if (jstreeId === 'treeFolder-img') {
           $('#sel_colorIcon').hide()
-          $('#bt_uploadImg').attr('data-path', realPath);
+          $('#bt_uploadImg').attr('data-path', realPath)
           for (var i in data) {
-            div += '<div class="divIconSel divImgSel">';
-            div += '<div class="cursor iconSel"><img class="img-responsive" src="' + realPath + data[i] + '"/></div>';
-            div += '<div class="iconDesc">' + capitalizeFirstLetter(data[i].substr(0, data[i].lastIndexOf('.')).substr(0, 15).split('_').join(' ')) + '</div>';
-            div += '<a class="btn btn-danger btn-xs bt_removeImg" data-realfilepath="' + realPath + data[i] + '"><i class="fas fa-trash"></i> {{Supprimer}}</a>';
-            div += '</div>';
+            div += '<div class="divIconSel divImgSel">'
+            div += '<div class="cursor iconSel"><img class="img-responsive" src="' + realPath + data[i] + '"/></div>'
+            div += '<div class="iconDesc">' + capitalizeFirstLetter(data[i].substr(0, data[i].lastIndexOf('.')).substr(0, 15).split('_').join(' ')) + '</div>'
+            div += '<a class="btn btn-danger btn-xs bt_removeImg" data-realfilepath="' + realPath + data[i] + '"><i class="fas fa-trash-alt"></i> {{Supprimer}}</a>'
+            div += '</div>'
           }
         } else if (jstreeId === 'treeFolder-bg') {
           $('#sel_colorIcon').hide()
           for (var i in data) {
-            div += '<div class="divIconSel divBgSel">';
-            div += '<div class="cursor iconSel"><img class="img-responsive" src="' + realPath + data[i] + '" data-filename="' + _path + data[i] + '"/></div>';
-            div += '<div class="iconDesc">' + capitalizeFirstLetter(data[i].substr(0, data[i].lastIndexOf('.')).split('_').join(' ')) + '</div>';
-            div += '</div>';
+            div += '<div class="divIconSel divBgSel">'
+            div += '<div class="cursor iconSel"><img class="img-responsive" src="' + realPath + data[i] + '" data-filename="' + _path + data[i] + '"/></div>'
+            div += '<div class="iconDesc">' + capitalizeFirstLetter(data[i].substr(0, data[i].lastIndexOf('.')).split('_').join(' ')) + '</div>'
+            div += '</div>'
           }
 
         }
@@ -262,14 +368,14 @@ sendVarToJS([
   }
 
   $('#mod_selectIcon ul.nav.nav-tabs li a').click(function() {
-    $.hideAlert();
+    $.hideAlert()
     var tabhref = $(this).attr('href')
     if (tabhref === '#tabicon') {
       $('#sel_colorIcon').show()
     } else {
       $('#sel_colorIcon').hide()
       if ($(tabhref + ' .div_treeFolder .jstree-clicked').length === 0) {
-        $(tabhref + ' .div_treeFolder').find('a:first').click();
+        $(tabhref + ' .div_treeFolder').find('a:first').click()
       }
     }
   })
@@ -281,7 +387,7 @@ sendVarToJS([
   if (jeephp2js.md_iconSelector_selectIcon != "0") {
     var iconClasses = jeephp2js.md_iconSelector_selectIcon.split('.')
     if (iconClasses[1].substr(0, 2) === 'fa') {
-      iconClasses[1] = 'font-awesome5';
+      iconClasses[1] = 'font-awesome5'
     } else if (iconClasses[1] === 'icon') {
       iconClasses[1] = iconClasses[2].split('-')[0]
     }
@@ -310,60 +416,52 @@ sendVarToJS([
 
   $('#bt_uploadImg').fileupload({
     add: function(e, data) {
-      let currentPath = $('#bt_uploadImg').attr('data-path');
-      data.url = 'core/ajax/jeedom.ajax.php?action=uploadImageIcon&filepath=' + currentPath;
-      data.submit();
+      let currentPath = $('#bt_uploadImg').attr('data-path')
+      data.url = 'core/ajax/jeedom.ajax.php?action=uploadImageIcon&filepath=' + currentPath
+      data.submit()
     },
     done: function(e, data) {
       if (data.result.state != 'ok') {
-        $('#div_iconSelectorAlert').showAlert({
+        $.fn.showAlert({
           message: data.result.result,
           level: 'danger'
-        });
-        return;
+        })
+        return
       }
-      $('.jstree-clicked').click();
-      $('#div_iconSelectorAlert').showAlert({
-        message: 'Fichier(s) ajouté(s) avec succès',
-        level: 'success'
-      });
+      $('.jstree-clicked').click()
     }
-  });
+  })
 
   $('#tabimg .div_imageGallery').off('click').on('click', '.bt_removeImg', function() {
-    $.hideAlert();
-    var filepath = $(this).attr('data-realfilepath');
-    bootbox.confirm('{{Êtes-vous sûr de vouloir supprimer cette image}} <span style="font-weight: bold ;">' + filepath + '</span> ?', function(result) {
+    $.hideAlert()
+    var filepath = $(this).attr('data-realfilepath')
+    bootbox.confirm('{{Êtes-vous sûr de vouloir supprimer cette image}} <strong>' + filepath + '</strong> ?', function(result) {
       if (result) {
         jeedom.removeImageIcon({
           filepath: filepath,
           error: function(error) {
-            $('#div_iconSelectorAlert').showAlert({
+            $.fn.showAlert({
               message: error.message,
               level: 'danger'
-            });
+            })
           },
-          success: function(data) {
-            $('.jstree-clicked').click();
-            $('#div_iconSelectorAlert').showAlert({
-              message: 'Fichier supprimé avec succès',
-              level: 'success'
-            });
+          success: function() {
+            $('.jstree-clicked').click()
           }
         })
       }
     })
-  });
+  })
 
   $('.div_imageGallery').on('click', '.divIconSel', function() {
-    $('.divIconSel').removeClass('iconSelected');
-    $(this).closest('.divIconSel').addClass('iconSelected');
+    $('.divIconSel').removeClass('iconSelected')
+    $(this).closest('.divIconSel').addClass('iconSelected')
   })
 
   $('.div_imageGallery').on('dblclick', '.divIconSel', function() {
-    $('.divIconSel').removeClass('iconSelected');
-    $(this).closest('.divIconSel').addClass('iconSelected');
-    $('#mod_selectIcon').dialog("option", "buttons")['Valider'].apply($('#mod_selectIcon'));
+    $('.divIconSel').removeClass('iconSelected')
+    $(this).closest('.divIconSel').addClass('iconSelected')
+    $('#mod_selectIcon').dialog("option", "buttons")['Valider'].apply($('#mod_selectIcon'))
   })
 
   setTimeout(function() {